+++ layers/graph_layer/graph_layer.py
@@ -1,6 +1,4 @@
-from layers.sp_utils import compute_diff
 import torch
-from torch.autograd import Function
 import torch.nn as nn
